Linkedin assistant/requirements.txt
flask>=3.0.0
python-telegram-bot[job-queue]==21.7
openai>=1.12.0,<2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0,<0.28.0
